Fix crash when fps are double values
Tried with a simple Lottie file and received an error about the fps being an invalid value. I researched and found out that the value was given with a lot of decimals and that was making trouble. Just rounded the value with no decimals and the error was fixed. I made a few tests and no content is loss.